# docs/swagger/swagger.yaml
basePath: /core/v1
definitions:
...
  lakukan-be_internal_modules_crm_leads_dto.CreateLeadRequest:
    properties:
      address:
        type: string
      assigned_to_company_user_id:
        type: string
      category:
        enum:
        - hot
        - warm
        - cold
        type: string
      city:
        maxLength: 100
        type: string
      contact_person:
        maxLength: 255
        type: string
      deal_value:
        minimum: 0
        type: number
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      next_follow_up_at:
        type: string
      notes:
        type: string
      phone:
        maxLength: 20
        type: string
      postal_code:
        maxLength: 10
        type: string
      province:
        maxLength: 100
        type: string
      source_id:
        type: string
      status:
        enum:
        - new
        - in_progress
        - follow_up
        - negotiation
        - won
        - lost
        type: string
    required:
    - category
    - name
    - phone
    - source_id
    type: object
  lakukan-be_internal_modules_crm_leads_dto.UpdateLeadRequest:
    properties:
      address:
        type: string
      assigned_to_company_user_id:
        type: string
      category:
        enum:
        - hot
        - warm
        - cold
        type: string
      city:
        maxLength: 100
        type: string
      contact_person:
        maxLength: 255
        type: string
      deal_value:
        minimum: 0
        type: number
      email:
        maxLength: 255
        type: string
      last_contact_at:
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      next_follow_up_at:
        type: string
      notes:
        type: string
      phone:
        maxLength: 20
        type: string
      postal_code:
        maxLength: 10
        type: string
      province:
        maxLength: 100
        type: string
      source_id:
        type: string
      status:
        enum:
        - new
        - in_progress
        - follow_up
        - negotiation
        - won
        - lost
        type: string
    type: object
...
  lakukan-be_internal_modules_crm_webhooks_dto.WAHAMeInfo:
    properties:
      id:
        type: string
      pushName:
        type: string
    type: object
  lakukan-be_internal_modules_crm_webhooks_dto.WAHAWebhookPayload:
    properties:
      engine:
        type: string
      environment:
        additionalProperties: true
        type: object
      event:
        type: string
      id:
        type: string
      me:
        $ref: '#/definitions/lakukan-be_internal_modules_crm_webhooks_dto.WAHAMeInfo'
      metadata:
        additionalProperties:
          type: string
        type: object
      payload:
        additionalProperties: true
        type: object
      session:
        type: string
      timestamp:
        type: integer
    type: object
  lakukan-be_internal_shared_response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/lakukan-be_internal_shared_response.Pagination'
    type: object
  lakukan-be_internal_shared_response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  lakukan-be_internal_shared_response.Response:
    properties:
      data: {}
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        type: string
      meta:
        $ref: '#/definitions/lakukan-be_internal_shared_response.Meta'
    type: object
host: localhost:8080
info:
  contact:
    email: support@lakukan.com
    name: API Support
    url: http://www.lakukan.com/support
  description: Backend API untuk Lakukan - Multi-tenant Business Management Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Lakukan API
  version: "1.0"
paths:
...
  /crm/v1/leads:
    get:
      consumes:
      - application/json
      description: Get list of leads with pagination and filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by name, phone, email, or contact person
        in: query
        name: search
        type: string
      - description: Filter by category (hot, warm, cold)
        in: query
        name: category
        type: string
      - description: Filter by status (new, in_progress, follow_up, negotiation, won,
          lost)
        in: query
        name: status
        type: string
      - description: Filter by source ID
        in: query
        name: source_id
        type: string
      - description: Filter by assigned user
        in: query
        name: assigned_to_company_user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get all leads
      tags:
      - CRM - Leads
    post:
      consumes:
      - application/json
      description: Create a new lead record
      parameters:
      - description: Lead data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_crm_leads_dto.CreateLeadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Create new lead
      tags:
      - CRM - Leads
  /crm/v1/leads/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a lead
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Delete lead
      tags:
      - CRM - Leads
    get:
      consumes:
      - application/json
      description: Get a single lead by ID
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get lead by ID
      tags:
      - CRM - Leads
    put:
      consumes:
      - application/json
      description: Update an existing lead
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      - description: Lead data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_crm_leads_dto.UpdateLeadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Update lead
      tags:
      - CRM - Leads
  /crm/v1/leads/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restore a soft-deleted lead
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Restore deleted lead
      tags:
      - CRM - Leads
...
  /crm/v1/webhooks/waha:
    post:
      consumes:
      - application/json
      description: Receive and process webhooks from WAHA (WhatsApp HTTP API)
      parameters:
      - description: Webhook request ID
        in: header
        name: X-Webhook-Request-Id
        type: string
      - description: Webhook timestamp
        in: header
        name: X-Webhook-Timestamp
        type: string
      - description: HMAC signature for verification
        in: header
        name: X-Webhook-Hmac
        type: string
      - description: Webhook payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_crm_webhooks_dto.WAHAWebhookPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      summary: Receive WAHA webhook
      tags:
      - CRM - Webhooks
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"